Refactor macros
