+++ src/interface.rs
@@ -1,3 +1,4 @@
 mod func_ptr;
 pub mod injector;
+mod macros;
 mod verifier;
